+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 25.4.3
+* ! Minor breaking change ! User properties will now be automatically saved under the following conditions:
+  * When an event is recorded
+  * During an internal timer tick
+  * When a session call made
+
+* Added a new function "UserDetails.Save()" for enqueuing cached user details manually. User details will now be saved upon with above triggers and manual "Save" call.
+
+* Mitigated an issue preventing recording multiple user details.
+
 ## 25.4.2
 * Added support for distinguishing macOS as a distinct OS value in metrics.